jobs/requirements.txt
Django==4.2
psycopg2-binary==2.9.9
Faker==33.0.0
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
